norm: here_document.c
