+++ Makefile
@@ -35,6 +35,7 @@ SRCS		=	$(SRCS_DIR)main.c	\
 				$(SRCS_DIR)parse_destroy_tree.c		\
 				$(SRCS_DIR)parse_error_handling.c	\
 				$(SRCS_DIR)here_document.c	\
+				$(SRCS_DIR)here_document_unlink.c	\
 				$(SRCS_DIR)traverse.c	\
 				$(SRCS_DIR)expansion_ctl1.c				\
 				$(SRCS_DIR)expansion_ctl2.c				\
