+++ src/here_document.c
@@ -69,6 +69,12 @@ static int	here_doc_action(char *filename, char *limiter)
 	return (0);
 }
 
+static void	switch_here_doc_norm(t_dict *env_dict, int fd)
+{
+	dict_modi_val_or_new_in_sort(env_dict, "?", ft_itoa(1));
+	func_guard(dup2(fd, 0), PROGRAM_NAME, "here_doc_traverse().");
+}
+
 static int	switch_here_doc(t_tree *tree_left, t_list **here_doc_list,
 		t_dict *env_dict)
 {
@@ -91,27 +97,10 @@ static int	switch_here_doc(t_tree *tree_left, t_list **here_doc_list,
 	free(limiter);
 	tree_left->left = filename;
 	if (g_signal)
-	{
-		dict_modi_val_or_new_in_sort(env_dict, "?", ft_itoa(1));
-		func_guard(dup2(fd, 0), PROGRAM_NAME, "here_doc_traverse().");
-	}
+		switch_here_doc_norm(env_dict, fd);
 	func_guard(close(fd), PROGRAM_NAME, "here_doc_traverse().");
 	set_signal(0);
 	return (g_signal);
-}
-
-void	unlink_here_doc_temp_file(t_list **here_doc_list)
-{
-	t_list	*node;
-
-	node = *here_doc_list;
-	while (node)
-	{
-		func_guard(unlink(node->content),
-			PROGRAM_NAME, "unlink_here_doc_temp_file().");
-		node = node->next;
-	}
-	ft_lstclear(here_doc_list, free);
 }
 
 int	here_doc_traverse(t_tree *tree, t_list **here_doc_list, t_dict *env_dict)
